requirements.txt
flask
pandas
requests
beautifulsoup4
scikit-learn
sentence-transformers
spacy
pdfplumber
python-docx
tabulate 
transformers